# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List review queue recommendations
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecommendationRecord"
    post:
      operationId: createRecommendation
      tags: [recommendations]
      summary: Create a pending recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /recommendations/{id}/status:
    post:
      operationId: changeRecommendationStatus
      tags: [recommendations]
      summary: Approve, reject, or revise a recommendation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationStatusChange"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    RecommendationCategory:
      type: string
      enum:
        - duplicate
        - team-pulse
        - sop-update
        - automation
        - mockup-prompt
        - market
        - mind-meld-handoff
        - company-record
        - sensitive
        - final-decision
        - external-intake
    RecommendationClassification:
      type: string
      enum:
        - documented-fact
        - user-update
        - ai-recommendation
        - draft-idea
        - pending-approval
        - approved-decision
        - sensitive
    RecommendationRiskLevel:
      type: string
      enum: [low, medium, high, critical]
    RecommendationApprovalRoute:
      type: string
      enum: [rose, carmen, both, none]
    RecommendationStatus:
      type: string
      enum: [pending, approved, rejected, needs-revision]
    RecommendationHistoryEntry:
      type: object
      required: [id, timestamp, actor, action]
      properties:
        id: { type: string }
        timestamp: { type: string }
        actor: { type: string }
        action: { type: string }
    RecommendationApprovals:
      type: object
      required: [rose, carmen]
      properties:
        rose: { type: boolean }
        carmen: { type: boolean }
    RecommendationRecord:
      type: object
      required:
        - id
        - source
        - category
        - recommendation
        - classification
        - risk
        - requiredApprover
        - status
        - approvals
        - history
        - createdAt
        - updatedAt
      properties:
        id: { type: integer }
        source: { type: string }
        category: { $ref: "#/components/schemas/RecommendationCategory" }
        recommendation: { type: string }
        classification: { $ref: "#/components/schemas/RecommendationClassification" }
        risk: { $ref: "#/components/schemas/RecommendationRiskLevel" }
        requiredApprover: { $ref: "#/components/schemas/RecommendationApprovalRoute" }
        status: { $ref: "#/components/schemas/RecommendationStatus" }
        approvals: { $ref: "#/components/schemas/RecommendationApprovals" }
        history:
          type: array
          items: { $ref: "#/components/schemas/RecommendationHistoryEntry" }
        createdById: { type: ["integer", "null"] }
        createdByName: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    RecommendationInput:
      type: object
      required: [source, category, recommendation, classification, risk, requiredApprover]
      properties:
        source: { type: string, minLength: 1 }
        category: { $ref: "#/components/schemas/RecommendationCategory" }
        recommendation: { type: string, minLength: 1 }
        classification: { $ref: "#/components/schemas/RecommendationClassification" }
        risk: { $ref: "#/components/schemas/RecommendationRiskLevel" }
        requiredApprover: { $ref: "#/components/schemas/RecommendationApprovalRoute" }
    RecommendationStatusChange:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/RecommendationStatus" }
...