# lib/api-zod/src/generated/types/recommendationApprovalRoute.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RecommendationApprovalRoute = typeof RecommendationApprovalRoute[keyof typeof RecommendationApprovalRoute];


export const RecommendationApprovalRoute = {
  rose: 'rose',
  carmen: 'carmen',
  both: 'both',
  none: 'none',
} as const;